Move coefficient proratisation to new file
