+++ openfisca_france/model/prelevements_obligatoires/prelevements_sociaux/cotisations_sociales/allegements.py
@@ -37,80 +37,6 @@ class assiette_allegement(Variable):
             )
 
 
-class coefficient_proratisation(Variable):
-    column = FloatCol
-    entity_class = Individus
-    label = u"Coefficient de proratisation pour le calcul du SMIC et du plafond de la Sécurité sociale"
-
-    def function(self, simulation, period):
-        # u"temps_plein",
-        # u"temps_partiel",
-        # u"forfait_heures_semaines",
-        # u"forfait_heures_mois",
-        # u"forfait_heures_annee",
-        # u"forfait_jours_annee",
-        contrat_de_travail = simulation.calculate('contrat_de_travail', period)
-        contrat_de_travail_debut = simulation.calculate('contrat_de_travail_debut', period)
-        contrat_de_travail_fin = simulation.calculate('contrat_de_travail_fin', period)
-        forfait_heures_remunerees_volume = simulation.calculate('forfait_heures_remunerees_volume', period)
-        forfait_jours_remuneres_volume = simulation.calculate('forfait_jours_remuneres_volume', period)
-        heures_duree_collective_entreprise = simulation.calculate('heures_duree_collective_entreprise', period)
-        heures_remunerees_volume = simulation.calculate('heures_remunerees_volume', period)
-        heures_non_remunerees_volume = simulation.calculate('heures_non_remunerees_volume', period)
-
-        # Décompte des jours en début et fin de contrat
-        # http://www.gestiondelapaie.com/flux-paie/?1029-la-bonne-premiere-paye
-
-        #  * Tous les calculs sont faits sur le mois *
-
-        # Méthode numpy de calcul des jours travaillés
-        # @holidays : jours feriés français
-        busday_count = partial(original_busday_count, holidays = holidays)
-
-        debut_mois = datetime64(period.start.offset('first-of', 'month'))
-        fin_mois = datetime64(period.start.offset('last-of', 'month')) + timedelta64(1, 'D') # busday ignores the last day
-
-        mois_incomplet = or_(contrat_de_travail_debut > debut_mois, contrat_de_travail_fin < fin_mois)
-        # jours travaillés ce mois-ci : intersection du contrat de travail et du mois en cours
-        jours_travailles_ce_mois = busday_count(
-            max_(contrat_de_travail_debut, debut_mois),
-            min_(contrat_de_travail_fin, fin_mois)
-            )
-
-        duree_legale_mensuelle = 35 * 52 / 12  # mensuelle_temps_plein
-        heures_temps_plein = (
-            (heures_duree_collective_entreprise == 0) * duree_legale_mensuelle + heures_duree_collective_entreprise
-            )
-        # heures remunerees avant conges sans soldes/ijss
-        heures_remunerees_volume = (
-            # Salariés à temps plein
-            (contrat_de_travail == 0) * (
-                heures_temps_plein * not_(mois_incomplet) +  # par défaut 151.67
-                jours_travailles_ce_mois * 7 * mois_incomplet  # TODO: 7 = heures / jours
-                ) +
-            # Salariés sans convention de forfait à temps partiel
-            (contrat_de_travail == 1) * heures_remunerees_volume
-            )
-
-        heures_realisees = heures_remunerees_volume - heures_non_remunerees_volume
-
-        coefficient = (
-            # Salariés à temps plein
-            (contrat_de_travail == 0) * heures_realisees / heures_temps_plein +
-            # Salariés à temps partiel : plafond proratisé en fonction du ratio durée travaillée / durée du temps plein
-            #   Salariés sans convention de forfait à temps partiel
-            (contrat_de_travail == 1) * heures_realisees / heures_temps_plein +
-            #   Salariés avec convention de forfait
-            #      Forfait en heures
-            (contrat_de_travail >= 2) * (contrat_de_travail <= 3) * (
-                forfait_heures_remunerees_volume / 45.7 * 52 / 12
-                ) +
-            #      Forfait en jours
-            (contrat_de_travail == 4) * forfait_jours_remuneres_volume / 218
-            )
-        return period, coefficient
-
-
 class credit_impot_competitivite_emploi(DatedVariable):
     column = FloatCol
     entity_class = Individus
